+++ src/imio/smartweb/core/behaviors/subsite.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from imio.smartweb.core.contents import IFolder
+from imio.smartweb.core.contents.cropping import SmartwebCroppingProvider
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.namedfile.field import NamedBlobImage
@@ -9,6 +10,15 @@ from zope import schema
 from zope.component import adapter
 from zope.interface import implementer
 from zope.interface import provider
+
+
+class SubsiteCroppingProvider(SmartwebCroppingProvider):
+    def get_scales(self, fieldname, request=None):
+        if fieldname == "logo":
+            # scale used for logo field
+            return ["preview"]
+        else:
+            return super(SubsiteCroppingProvider, self).get_scales(fieldname, request)
 
 
 @provider(IFormFieldProvider)