+++ src/imio/smartweb/core/contents/folder/content.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from collective.instancebehavior.interfaces import IInstanceBehaviorAssignableContent
-from imio.smartweb.common.adapters import BaseCroppingProvider
 from imio.smartweb.core.contents import IDefaultPages
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api
@@ -15,15 +14,6 @@ from zope import schema
 from zope.interface import implementer
 from zope.interface import alsoProvides
 from zope.interface import noLongerProvides
-
-
-class FolderCroppingProvider(BaseCroppingProvider):
-    def get_scales(self, fieldname, request=None):
-        if fieldname == "banner":
-            # scale used for banner field
-            return ["banner"]
-        else:
-            return super(FolderCroppingProvider, self).get_scales(fieldname, request)
 
 
 class IFolder(model.Schema):