+++ src/imio/smartweb/core/contents/sections/base.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from imio.smartweb.common.adapters import BaseCroppingProvider
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone.autoform.directives import write_permission
 from plone.dexterity.content import Container
@@ -8,15 +7,6 @@ from plone.namedfile.field import NamedBlobImage
 from plone.supermodel import model
 from zope import schema
 from zope.interface import implementer
-
-
-class SectionCroppingProvider(BaseCroppingProvider):
-    def get_scales(self, fieldname, request=None):
-        if fieldname == "background_image":
-            # scale used for background_image field
-            return ["affiche"]
-        else:
-            return super(SectionCroppingProvider, self).get_scales(fieldname, request)
 
 
 class ISection(model.Schema):